pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "contract-testing"
version = "1.0.0"
description = "Contract testing system for API validation and breaking change detection"
requires-python = ">=3.10"
dependencies = [
    "jsonschema==4.26.0",
    "openapi-core==0.23.0",
    "prance==25.4.3.0",
    "openapi-spec-validator==0.8.4",
    "deepdiff==8.6.1",
    "genson>=1.3.0",
    "playwright>=1.49.0",
    "pytest>=8.0",
    "pytest-html>=4.0",
    "pytest-playwright>=0.6.2",
    "pyyaml>=6.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
markers = [
    "live_api: tests that call live external APIs (may be slow or flaky)",
    "diff: schema diff / breaking-change detection tests",
]
addopts = "-v --tb=short --html=reports/report.html --self-contained-html"

[tool.setuptools.packages.find]
include = ["src*"]